[BugFix][Embedded]fix lineHeight error
use the wrong lineHeight type for LayoutInElement

issue:m-6787660024

AutoLand:release/3.5

AutoSubmit:true

+++ platform/android/lynx_android/src/main/java/com/lynx/tasm/behavior/shadow/text/TextMeasurer.java
@@ -169,7 +169,7 @@ public class TextMeasurer {
 
           break;
         case kTextPropLineHeight:
-          int lineHeight = iterator.next().getInt();
+          float lineHeight = (float) (iterator.next().getDouble());
           if (!isParagraph) {
             Log.w("TextMeasurer", "line-height should be set to paragraph");
             continue;
